@@ -1,11 +1,5 @@
 <script>
-	import { createEventDispatcher } from 'svelte';
-	export let route;
-	const dispatch = createEventDispatcher();
-
-	function handleClick() {
-		dispatch('routeClick', { route });
-	}
+	let { handleModalRouteClick, route } = $props();
 
 	function getDisplayRouteName() {
 		if (route.shortName && route.longName) {
@@ -21,7 +15,7 @@
 <button
 	type="button"
 	class="route-item flex w-full items-center justify-between border-b border-gray-200 bg-[#f9f9f9] p-4 text-left hover:bg-[#e9e9e9] focus:outline-none dark:border-[#313135] dark:bg-[#1c1c1c] dark:text-white dark:hover:bg-[#363636]"
-	on:click={handleClick}
+	onclick={() => handleModalRouteClick(route)}
 >
 	<div class="text-lg font-semibold" style="color: #{route.color}">
 		{getDisplayRouteName(route)}
