@@ -2,24 +2,25 @@
 	import { setContext } from 'svelte';
 	import { writable, derived } from 'svelte/store';
 
-	// Props to track all possible item IDs
-	export let items = [];
-
 	// Store for animation state
 	const skipAnimation = writable(false);
 
 	// Create a store to track multiple active items using a Set
 	const activeItems = writable(new Set());
 
-	// Create dispatch for activeChanged event
-	import { createEventDispatcher } from 'svelte';
-	const dispatch = createEventDispatcher();
+	/**
+	 * @typedef {Object} Props
+	 * @property {any} [items] - Props to track all possible item IDs
+	 * @property {import('svelte').Snippet} [children]
+	 */
+
+	/** @type {Props} */
+	let { items = $bindable([]), children } = $props();
 
 	// Methods to open/close all items
 	export const openAll = (animate = true) => {
 		skipAnimation.set(!animate);
 		activeItems.set(new Set(items));
-		dispatch('openAll', { animated: animate });
 		// Reset skip animation after a short delay
 		setTimeout(() => skipAnimation.set(false), 0);
 	};
@@ -27,15 +28,9 @@
 	export const closeAll = (animate = true) => {
 		skipAnimation.set(!animate);
 		activeItems.set(new Set());
-		dispatch('closeAll', { animated: animate });
 		// Reset skip animation after a short delay
 		setTimeout(() => skipAnimation.set(false), 0);
 	};
-
-	// Watch for changes to activeItems and dispatch event
-	$: {
-		dispatch('activeChanged', { activeItems: Array.from($activeItems) });
-	}
 
 	// Provide context for child AccordionItems and expose methods
 	setContext('accordion', {
@@ -70,5 +65,5 @@
 <div
 	class="divide-y divide-gray-200 border-y border-gray-200 dark:divide-gray-700 dark:border-gray-700"
 >
-	<slot />
+	{@render children?.()}
 </div>
