@@ -2,15 +2,14 @@
 	import { fly } from 'svelte/transition';
 	import ThemeSwitcher from '$lib/ThemeSwitch/ThemeSwitcher.svelte';
 
-	export let headerLinks = {};
-	export let closeMenu;
+	let { headerLinks = {}, closeMenu } = $props();
 </script>
 
 <div
 	class="fixed inset-0 z-50 flex flex-col items-center justify-center space-y-6 bg-white p-4 dark:bg-black"
 	transition:fly={{ x: 1000, duration: 300 }}
 >
-	<button on:click={closeMenu}>
+	<button onclick={closeMenu} aria-label="Close Menu">
 		<svg
 			class="close-icon h-6 w-6 text-gray-900 dark:text-white"
 			fill="none"
@@ -28,7 +27,7 @@
 			<a
 				href={value}
 				class="block text-xl font-semibold text-gray-900 dark:text-white"
-				on:click={closeMenu}>{key}</a
+				onclick={closeMenu}>{key}</a
 			>
 		{/each}
 	</div>
