@@ -1,26 +1,24 @@
 <script>
 	import { debounce } from '$lib/utils';
-	import { createEventDispatcher, onMount } from 'svelte';
+	import { onMount } from 'svelte';
 	import TripPlanSearchField from './TripPlanSearchField.svelte';
 	import { error } from '@sveltejs/kit';
 	import { browser } from '$app/environment';
 	import { t } from 'svelte-i18n';
-	export let mapProvider;
+	let { handleTripPlan, mapProvider } = $props();
 
-	let fromPlace = '';
-	let toPlace = '';
-	let fromResults = [];
-	let toResults = [];
-	let selectedFrom = null;
-	let selectedTo = null;
-	let isLoadingFrom = false;
-	let isLoadingTo = false;
+	let fromPlace = $state('');
+	let toPlace = $state('');
+	let fromResults = $state([]);
+	let toResults = $state([]);
+	let selectedFrom = $state(null);
+	let selectedTo = $state(null);
+	let isLoadingFrom = $state(false);
+	let isLoadingTo = $state(false);
 	let fromMarker;
 	let toMarker;
-	let loading = false;
+	let loading = $state(false);
 	let lockSelectLocation = false;
-
-	const dispatch = createEventDispatcher();
 
 	async function fetchAutocompleteResults(query) {
 		const response = await fetch(
@@ -149,7 +147,12 @@
 			const data = await fetchTripPlan(selectedFrom, selectedTo);
 
 			if (data) {
-				dispatch('tripPlanned', { data, fromMarker, toMarker });
+				const tripPlanData = {
+					data,
+					fromMarker,
+					toMarker
+				};
+				handleTripPlan(tripPlanData);
 			}
 		} finally {
 			loading = false;
@@ -188,7 +191,7 @@
 	/>
 
 	<button
-		on:click={planTrip}
+		onclick={planTrip}
 		class="mt-4 flex w-full items-center justify-center rounded-md bg-green-500 py-2 text-white shadow-md
            transition-colors
            hover:bg-green-600
