@@ -12,30 +12,19 @@
 	import { isLoading } from 'svelte-i18n';
 	import { t } from 'svelte-i18n';
 
-	let selectedDate = new Date();
-	let prevSelectedDate = null;
+	let selectedDate = $state(new Date());
+	let prevSelectedDate = $state(null);
+	let emptySchedules = $state(false);
+	let schedules = $state([]);
+	let stopName = $state('');
+	let stopId = $state('');
+	let stopDirection = $state('');
+	let loading = $state(true);
+	let accordionComponent = $state();
+
 	let schedulesMap = new Map();
 	let routeReference = new Map();
-	let schedules = [];
-	let stopName = '';
-	let stopId = '';
-	let stopDirection = '';
-	let loading = true;
-	let emptySchedules = false;
 	let currentDate = new Date();
-	let accordionComponent;
-
-	$: stopId = $page.params.stopID;
-
-	$: if (selectedDate && selectedDate !== prevSelectedDate) {
-		const formattedDate = selectedDate.toISOString().split('T')[0];
-		prevSelectedDate = selectedDate;
-
-		// we get an error if we try to fetch data on the server
-		if (typeof window !== 'undefined') {
-			fetchScheduleForStop(stopId, formattedDate);
-		}
-	}
 
 	async function fetchScheduleForStop(stopId, date) {
 		try {
@@ -121,6 +110,20 @@
 		await fetchScheduleForStop(stopId, formattedDate);
 		accordionComponent.openAll(false);
 	});
+
+	stopId = $page.params.stopID;
+
+	$effect(() => {
+		if (selectedDate && selectedDate !== prevSelectedDate) {
+			const formattedDate = selectedDate.toISOString().split('T')[0];
+			prevSelectedDate = selectedDate;
+
+			// we get an error if we try to fetch data on the server
+			if (typeof window !== 'undefined') {
+				fetchScheduleForStop(stopId, formattedDate);
+			}
+		}
+	});
 </script>
 
 <svelte:head>
@@ -145,10 +148,10 @@
 					</div>
 
 					<div class="flex-1 text-right">
-						<button class="button" on:click={() => accordionComponent.openAll()}>
+						<button class="button" onclick={() => accordionComponent.openAll()}>
 							{$t('schedule_for_stop.show_all_routes')}
 						</button>
-						<button class="button" on:click={() => accordionComponent.closeAll()}>
+						<button class="button" onclick={() => accordionComponent.closeAll()}>
 							{$t('schedule_for_stop.collapse_all_routes')}
 						</button>
 					</div>
@@ -163,7 +166,9 @@
 						<Accordion bind:this={accordionComponent}>
 							{#each schedules as schedule}
 								<AccordionItem>
-									<span slot="header">{schedule.tripHeadsign}</span>
+									{#snippet header()}
+										<span>{schedule.tripHeadsign}</span>
+									{/snippet}
 									<RouteScheduleTable {schedule} />
 								</AccordionItem>
 							{/each}
