@@ -8,6 +8,7 @@ import PopupContent from '$components/map/PopupContent.svelte';
 import { createVehicleIconSvg } from '$lib/MapHelpers/generateVehicleIcon';
 import VehiclePopupContent from '$components/map/VehiclePopupContent.svelte';
 import TripPlanPinMarker from '$components/trip-planner/tripPlanPinMarker.svelte';
+import { mount, unmount } from 'svelte';
 
 export default class OpenStreetMapProvider {
 	constructor() {
@@ -64,13 +65,16 @@ export default class OpenStreetMapProvider {
 
 		const container = document.createElement('div');
 
-		const stopMarkerComponent = new StopMarker({
+		const props = $state({
+			stop: options.stop,
+			icon: options.icon || faBus,
+			onClick: options.onClick || (() => {}),
+			isHighlighted: false
+		});
+
+		mount(StopMarker, {
 			target: container,
-			props: {
-				stop: options.stop,
-				icon: options.icon || faBus,
-				onClick: options.onClick || (() => {})
-			}
+			props
 		});
 
 		const customIcon = this.L.divIcon({
@@ -83,7 +87,7 @@ export default class OpenStreetMapProvider {
 			icon: customIcon
 		}).addTo(this.map);
 
-		marker.stopMarkerComponent = stopMarkerComponent;
+		marker.props = props;
 
 		this.markersMap.set(options.stop.id, marker);
 
@@ -95,7 +99,7 @@ export default class OpenStreetMapProvider {
 
 		const container = document.createElement('div');
 
-		new TripPlanPinMarker({
+		mount(TripPlanPinMarker, {
 			target: container,
 			props: {
 				text: text
@@ -125,13 +129,16 @@ export default class OpenStreetMapProvider {
 	highlightMarker(stopId) {
 		const marker = this.markersMap.get(stopId);
 		if (!marker) return;
-		marker.stopMarkerComponent.$set({ isHighlighted: true });
+
+		// Update the reactive props (linked via $state)
+		marker.props.isHighlighted = true;
 	}
 
 	unHighlightMarker(stopId) {
 		const marker = this.markersMap.get(stopId);
 		if (!marker) return;
-		marker.stopMarkerComponent.$set({ isHighlighted: false });
+
+		marker.props.isHighlighted = false;
 	}
 
 	addStopMarker(stop, stopTime = null) {
@@ -152,12 +159,12 @@ export default class OpenStreetMapProvider {
 			}
 
 			if (this.popupContentComponent) {
-				this.popupContentComponent.$destroy();
+				unmount(this.popupContentComponent);
 			}
 
 			const popupContainer = document.createElement('div');
 
-			this.popupContentComponent = new PopupContent({
+			this.popupContentComponent = mount(PopupContent, {
 				target: popupContainer,
 				props: {
 					stopName: stop.name,
@@ -228,7 +235,7 @@ export default class OpenStreetMapProvider {
 		marker.on('popupopen', () => {
 			const popupContainer = document.createElement('div');
 
-			marker.popupComponent = new VehiclePopupContent({
+			marker.popupComponent = mount(VehiclePopupContent, {
 				target: popupContainer,
 				props: marker.vehicleData
 			});
@@ -238,7 +245,7 @@ export default class OpenStreetMapProvider {
 
 		marker.on('popupclose', () => {
 			if (marker.popupComponent) {
-				marker.popupComponent.$destroy();
+				unmount(marker.popupComponent);
 				marker.popupComponent = null;
 			}
 		});
